+++ tests/client/CallableOperation/construct_params_test.py
@@ -3,31 +3,8 @@ import pytest
 
 from bravado_core.exception import SwaggerMappingError
 from bravado_core.operation import Operation
-from bravado_core.spec import Spec
 
 from bravado.client import CallableOperation
-
-
-@pytest.fixture
-def request_dict():
-    return {}
-
-
-@pytest.fixture
-def getPetById_spec(petstore_dict):
-    return petstore_dict['paths']['/pet/{petId}']['get']
-
-
-@pytest.fixture
-def minimal_swagger_spec(getPetById_spec):
-    spec_dict = {
-        'paths': {
-            '/pet/{petId}': {
-                'get': getPetById_spec
-            }
-        }
-    }
-    return Spec(spec_dict)
 
 
 def test_simple(minimal_swagger_spec, getPetById_spec, request_dict):